Merge dev into master

+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Services/Methodologies/MethodologyImageService.cs
@@ -63,10 +63,18 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Services.Methodologie
                 {
                     await files.ForEachAsync(async file =>
                     {
-                        // TODO EES-1263 Methodology amendments - Check if file is linked to other methodolgies before deleting blob
-                        await _blobStorageService.DeleteBlob(PrivateMethodologyFiles, file.Path());
+                        var methodologyLinks = await _methodologyFileRepository.GetMethodologyLinksToFile(file.Id);
+
                         await _methodologyFileRepository.Delete(methodologyId, file.Id);
-                        await _fileRepository.Delete(file.Id);
+
+                        // If this Methodology is the only Methodology that is referencing this Blob and File, it
+                        // can be deleted from Blob Storage and the File table.  Otherwise preserve them for the other
+                        // Methodology versions that are still using them. 
+                        if (methodologyLinks.Count == 1 && methodologyLinks[0].MethodologyId == methodologyId)
+                        {
+                            await _blobStorageService.DeleteBlob(PrivateMethodologyFiles, file.Path());
+                            await _fileRepository.Delete(file.Id);
+                        }
                     });
                 });
         }
